[ Coverity ] Fix Coverity Issues
This PR incluide Fixs of Coverity Issues.

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

+++ nntrainer/src/databuffer_file.cpp
@@ -191,7 +191,7 @@ void DataBufferFromDataFile::updateData(BufferType type) {
     break;
   }
 
-  unsigned int I;
+  uint64_t I;
   std::vector<unsigned int> mark;
   mark.resize(max_size);
   file.clear();
